Switch to wyhash from CRC for capture tile diff
Originally suggested by @trishume at #167.

+++ module/rdpClientCon.h
@@ -123,7 +123,7 @@ struct _rdpClientCon
     RegionPtr dirtyRegion;
 
     int num_rfx_crcs_alloc[16];
-    int *rfx_crcs[16];
+    uint64_t *rfx_crcs[16];
 
     /* true = skip drawing */
     int suppress_output;
